Adjusted the layout and removed redundant styles. Renamed and re-organised some files.
