+++ src/assets/main.css
@@ -3,7 +3,7 @@
 #app {
   max-width: 1280px;
   margin: 0 auto;
-  padding: 1rem;
+  min-height: 100vh;
   background-color: blueviolet;
 }
