+++ src/pages/HomePage.vue
@@ -1,79 +1,20 @@
 <template>
-  <AppStructure>
+  <PageLayout>
     <template #header>
       <HeaderView />
     </template>
     <HomeView />
     <template #nav>
-      <HomeNavView />
+      <HomeNav />
     </template>
-  </AppStructure>
+  </PageLayout>
 </template>
 
 <script setup lang="ts">
-import AppStructure from "../components/AppStructure.vue";
-import HeaderView from "../views/HeaderView.vue";
+import PageLayout from "../components/PageLayout.vue";
+import HeaderView from "../views/headers/HeaderView.vue";
 import HomeView from "../views/HomeView.vue";
-import HomeNavView from "../views/HomeNavView.vue";
+import HomeNav from "../views/navs/HomeNav.vue";
 </script>
 
-<style scoped>
-header {
-  line-height: 1.3;
-  max-height: 100vh;
-  background-color: aqua;
-}
-
-nav {
-  width: 100%;
-  font-size: 12px;
-  text-align: center;
-  /* margin-top: 2rem; */
-  background-color: yellow;
-}
-
-nav a.router-link-exact-active {
-  color: var(--color-text);
-}
-
-nav a.router-link-exact-active:hover {
-  background-color: transparent;
-}
-
-nav a {
-  display: inline-block;
-  padding: 0 1rem;
-  border-left: 1px solid var(--color-border);
-}
-
-nav a:first-of-type {
-  border: 0;
-}
-
-@media (min-width: 1024px) {
-  header {
-    display: flex;
-    place-items: center;
-    padding-right: calc(var(--section-gap) / 2);
-  }
-
-  .logo {
-    margin: 0 2rem 0 0;
-  }
-
-  header .wrapper {
-    display: flex;
-    place-items: flex-start;
-    flex-wrap: wrap;
-  }
-
-  nav {
-    text-align: left;
-    margin-left: -1rem;
-    font-size: 1rem;
-
-    padding: 1rem 0;
-    margin-top: 1rem;
-  }
-}
-</style>
+<style scoped></style>
