+++ src/views/AboutView.vue
@@ -1,16 +1,7 @@
 <template>
-  <div class="about">
+  <div>
     <h1>This is an about page</h1>
   </div>
 </template>
 
-<style>
-@media (min-width: 1024px) {
-  .about {
-    min-height: 100vh;
-    display: flex;
-    align-items: center;
-    background-color: chocolate;
-  }
-}
-</style>
+<style scoped></style>
